R1/requirements.txt
# requirements.txt
psycopg2-binary
sqlalchemy